+++ src/modules/chrome.js
@@ -3,3 +3,8 @@ export async function getCurrentTab() {
   let [tab] = await chrome.tabs.query(queryOptions);
   return tab;
 };
+
+export function getExtensionManifest() {
+  const manifest = chrome.runtime.getManifest();
+  return manifest;
+}